requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
jinja2
python-dotenv
stripe
itsdangerous
groq